cloud_function/requirements.txt
google-api-python-client>=1.7.8,<2
google-cloud-aiplatform
kfp==2.3.0
functions-framework==3.*